# scrumX/backend/internal/apidocs/openapi.yaml
openapi: 3.0.3
info:
  title: ScrumX API
  version: 1.1.0
  description: |
    ScrumX REST API for issue/project/release/ITSM/integration workflows.
    This contract is intended for third-party integrations.
servers:
  - url: /api/v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    orgHeader:
      type: apiKey
      in: header
      name: X-Org-ID
  schemas:
...
    IntegrationCredentialPayload:
      type: object
      required: [credentials]
      properties:
        credentials:
          type: object
          additionalProperties: true
...
security:
  - bearerAuth: []
    orgHeader: []
paths:
...
  /integrations/github:
    post:
      summary: Upsert GitHub integration credentials (encrypted at rest)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IntegrationCredentialPayload"
      responses:
        "200":
          description: Upserted
  /integrations/jira:
    post:
      summary: Upsert Jira integration credentials (encrypted at rest)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IntegrationCredentialPayload"
      responses:
        "200":
          description: Upserted
  /integrations/{provider}/test:
    post:
      summary: Validate/decrypt stored integration config for provider
      parameters:
        - in: path
          name: provider
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Integration config is valid
...