# scrumX/backend/internal/integrations/handler.go
package integrations

import (
	"database/sql"
	"encoding/json"
	"errors"
	"strings"
	"time"

	"github.com/atharshah1/scrum/scrumX/backend/internal/authz"
	"github.com/atharshah1/scrum/scrumX/backend/pkg/middleware"
	"github.com/atharshah1/scrum/scrumX/backend/pkg/utils"
	"github.com/gofiber/fiber/v2"
	"github.com/google/uuid"
)

type Handler struct {
	db     *sql.DB
	authz  *authz.Service
	crypto *credentialCipher
}

func NewHandler(db *sql.DB, authzService *authz.Service, encryptionKey string) (*Handler, error) {
	crypto, err := newCredentialCipher(encryptionKey)
	if err != nil {
		return nil, err
	}
	return &Handler{db: db, authz: authzService, crypto: crypto}, nil
}

func (h *Handler) RegisterRoutes(api fiber.Router) {
	r := api.Group("/integrations")
	r.Get("/", h.list)
	r.Post("/github", h.upsertGitHub)
	r.Post("/jira", h.upsertJira)
	r.Post("/slack", h.upsertSlack)
	r.Post("/cicd", h.upsertCICD)
	r.Post("/deployments", h.upsertDeployments)
	r.Post("/:provider/test", h.testProvider)
}

func (h *Handler) list(c *fiber.Ctx) error {
	orgID, ok := middleware.MustOrgID(c)
	if !ok {
		return utils.JSONError(c, fiber.StatusBadRequest, "missing org context")
	}
	userID, ok := middleware.MustUserID(c)
	if !ok {
		return utils.JSONError(c, fiber.StatusUnauthorized, "missing user context")
	}
	if _, err := h.authz.RequireOrgMember(c.Context(), orgID, userID); err != nil {
		return utils.JSONError(c, fiber.StatusForbidden, "forbidden")
	}
	rows, err := h.db.QueryContext(c.Context(), `SELECT DISTINCT ON (provider) id, provider, credentials, created_at
FROM integrations
WHERE org_id=$1
ORDER BY provider, created_at DESC`, orgID)
	if err != nil {
		return utils.JSONError(c, fiber.StatusInternalServerError, err.Error())
	}
	defer rows.Close()
	out := make([]fiber.Map, 0, 8)
	for rows.Next() {
		var (
			id             uuid.UUID
			provider       string
			credentialsRaw []byte
			createdAt      time.Time
		)
		if err := rows.Scan(&id, &provider, &credentialsRaw, &createdAt); err != nil {
			return utils.JSONError(c, fiber.StatusInternalServerError, err.Error())
		}
		_, keys, err := h.crypto.Decrypt(credentialsRaw)
		if err != nil {
			return utils.JSONError(c, fiber.StatusInternalServerError, "unable to decrypt stored credentials")
		}
		out = append(out, fiber.Map{
			"id":              id,
			"provider":        provider,
			"configured":      len(keys) > 0,
			"credential_keys": keys,
			"created_at":      createdAt,
		})
	}
	return utils.JSONSuccess(c, fiber.StatusOK, out)
}

func (h *Handler) upsertGitHub(c *fiber.Ctx) error      { return h.upsertProvider(c, "github") }
func (h *Handler) upsertJira(c *fiber.Ctx) error        { return h.upsertProvider(c, "jira") }
func (h *Handler) upsertSlack(c *fiber.Ctx) error       { return h.upsertProvider(c, "slack") }
func (h *Handler) upsertCICD(c *fiber.Ctx) error        { return h.upsertProvider(c, "cicd") }
func (h *Handler) upsertDeployments(c *fiber.Ctx) error { return h.upsertProvider(c, "deployments") }

func (h *Handler) upsertProvider(c *fiber.Ctx, provider string) error {
	orgID, ok := middleware.MustOrgID(c)
	if !ok {
		return utils.JSONError(c, fiber.StatusBadRequest, "missing org context")
	}
	userID, ok := middleware.MustUserID(c)
	if !ok {
		return utils.JSONError(c, fiber.StatusUnauthorized, "missing user context")
	}
	role, err := h.authz.RequireOrgMember(c.Context(), orgID, userID)
	if err != nil || !authz.CanWrite(role) {
		return utils.JSONError(c, fiber.StatusForbidden, "forbidden")
	}
	var payload struct {
		Credentials map[string]any `json:"credentials"`
	}
	if err := c.BodyParser(&payload); err != nil || payload.Credentials == nil {
		return utils.JSONError(c, fiber.StatusBadRequest, "credentials are required")
	}
	if err := validateProviderCredentials(provider, payload.Credentials); err != nil {
		return utils.JSONError(c, fiber.StatusBadRequest, err.Error())
	}
	keys := mapKeys(payload.Credentials)
	credentialsRaw, err := json.Marshal(payload.Credentials)
	if err != nil {
		return utils.JSONError(c, fiber.StatusBadRequest, "invalid credentials")
	}
	encryptedCredentials, err := h.crypto.Encrypt(credentialsRaw, keys)
	if err != nil {
		return utils.JSONError(c, fiber.StatusInternalServerError, "failed to encrypt credentials")
	}
	tx, err := h.db.BeginTx(c.Context(), nil)
	if err != nil {
		return utils.JSONError(c, fiber.StatusInternalServerError, err.Error())
	}
	defer tx.Rollback()
	var integrationID uuid.UUID
	err = tx.QueryRowContext(c.Context(), `SELECT id FROM integrations WHERE org_id=$1 AND provider=$2 ORDER BY created_at DESC LIMIT 1`,
		orgID, provider).Scan(&integrationID)
	switch err {
	case nil:
		if _, err := tx.ExecContext(c.Context(), `UPDATE integrations SET credentials=$3 WHERE id=$1 AND org_id=$2`,
			integrationID, orgID, encryptedCredentials); err != nil {
			return utils.JSONError(c, fiber.StatusInternalServerError, err.Error())
		}
	case sql.ErrNoRows:
		integrationID = uuid.New()
		if _, err := tx.ExecContext(c.Context(), `INSERT INTO integrations (id, org_id, provider, credentials) VALUES ($1,$2,$3,$4)`,
			integrationID, orgID, provider, encryptedCredentials); err != nil {
			return utils.JSONError(c, fiber.StatusBadRequest, err.Error())
		}
	default:
		return utils.JSONError(c, fiber.StatusInternalServerError, err.Error())
	}
	if err := tx.Commit(); err != nil {
		return utils.JSONError(c, fiber.StatusInternalServerError, err.Error())
	}
	return utils.JSONSuccess(c, fiber.StatusOK, fiber.Map{
		"id":              integrationID,
		"provider":        provider,
		"configured":      true,
		"credential_keys": keys,
	})
}

func (h *Handler) testProvider(c *fiber.Ctx) error {
	orgID, ok := middleware.MustOrgID(c)
	if !ok {
		return utils.JSONError(c, fiber.StatusBadRequest, "missing org context")
	}
	userID, ok := middleware.MustUserID(c)
	if !ok {
		return utils.JSONError(c, fiber.StatusUnauthorized, "missing user context")
	}
	role, err := h.authz.RequireOrgMember(c.Context(), orgID, userID)
	if err != nil || !authz.CanWrite(role) {
		return utils.JSONError(c, fiber.StatusForbidden, "forbidden")
	}
	provider := strings.ToLower(strings.TrimSpace(c.Params("provider")))
	if provider == "" {
		return utils.JSONError(c, fiber.StatusBadRequest, "provider is required")
	}
	var credentialsRaw []byte
	if err := h.db.QueryRowContext(c.Context(), `SELECT credentials FROM integrations WHERE org_id=$1 AND provider=$2 ORDER BY created_at DESC LIMIT 1`,
		orgID, provider).Scan(&credentialsRaw); err != nil {
		if errors.Is(err, sql.ErrNoRows) {
			return utils.JSONError(c, fiber.StatusNotFound, "integration is not configured")
		}
		return utils.JSONError(c, fiber.StatusInternalServerError, err.Error())
	}
	decrypted, keys, err := h.crypto.Decrypt(credentialsRaw)
	if err != nil {
		return utils.JSONError(c, fiber.StatusInternalServerError, "unable to decrypt credentials")
	}
	decoded := map[string]any{}
	if err := json.Unmarshal(decrypted, &decoded); err != nil {
		return utils.JSONError(c, fiber.StatusInternalServerError, "stored credentials are invalid")
	}
	if err := validateProviderCredentials(provider, decoded); err != nil {
		return utils.JSONError(c, fiber.StatusBadRequest, err.Error())
	}
	return utils.JSONSuccess(c, fiber.StatusOK, fiber.Map{
		"provider":        provider,
		"configured":      true,
		"credential_keys": keys,
		"tested_at":       time.Now().UTC(),
	})
}

func mapKeys(decoded map[string]any) []string {
	keys := make([]string, 0, len(decoded))
	for k := range decoded {
		keys = append(keys, strings.TrimSpace(k))
	}
	return keys
}

func validateProviderCredentials(provider string, credentials map[string]any) error {
	requiredByProvider := map[string][]string{
		"github":      {"token"},
		"jira":        {"base_url", "email", "api_token"},
		"slack":       {"bot_token"},
		"cicd":        {"base_url", "token"},
		"deployments": {"base_url", "token"},
	}
	required, ok := requiredByProvider[provider]
	if !ok {
		return errors.New("unsupported integration provider")
	}
	for _, field := range required {
		raw, exists := credentials[field]
		if !exists || strings.TrimSpace(asString(raw)) == "" {
			return errors.New("missing required credential: " + field)
		}
	}
	return nil
}

func asString(value any) string {
	if value == nil {
		return ""
	}
	if v, ok := value.(string); ok {
		return v
	}
	return ""
}